+++ src/com/cloudogu/ces/dogubuildlib/EcoSystem.groovy
@@ -170,6 +170,37 @@ class EcoSystem {
         }
     }
 
+    void runDoguUpgradeTest(Map params, String doguName, String cypressImage = "cypress/included:8.7.0") {
+        stage('Upgrade dogu') {
+            // Remove new dogu that has been built and tested above
+            this.purgeDogu(doguName)
+
+            if (params.OldDoguVersionForUpgradeTest != '' && !params.OldDoguVersionForUpgradeTest.contains('v')){
+                println "Installing user defined version of dogu: " + params.OldDoguVersionForUpgradeTest
+                this.installDogu("official/" + doguName + " " + params.OldDoguVersionForUpgradeTest)
+            } else {
+                println "Installing latest released version of dogu..."
+                this.installDogu("official/" + doguName)
+            }
+            this.startDogu(doguName)
+            this.waitForDogu(doguName)
+            this.upgradeDogu(this)
+
+            // Wait for upgraded dogu to get healthy
+            this.waitForDogu(doguName)
+            waitUntilAvailable(doguName)
+        }
+
+        stage('Integration Tests - After Upgrade') {
+            // Run integration tests again to verify that the upgrade was successful
+            ecoSystem.runCypressIntegrationTests([
+                    cypressImage     : cypressImage,
+                    enableVideo      : params.EnableVideoRecording,
+                    enableScreenshots: params.EnableScreenshotRecording
+            ])
+        }
+    }
+
     void runYarnIntegrationTests(int timeoutInMinutes, String nodeImage, ArrayList<String> additionalArgs = [], boolean enableVideoRecording = false) {
         script.sh 'rm -f integrationTests/it-results.xml'
         def additionalContainerRunArgs = "${parseAdditionalIntegrationTestArgs(additionalArgs)} "